Moved transaction list to a separate component
